[CVP-1897] Generate mock prow artifacts when running external Operator Bundle Tests
When operator bundle deployment tests are run through DPTP, two files are generated that the upstream `optional-operators-subscribe` role does not generate.
We will need to generate these files as a part of the `optional-operators-subscribe` role.
The finished.json file should contain the following fields:
timestamp: use the already-existing `start_time` field
passed: boolean true or false
result: Omit. The result field has been deprecated in favor of passed
revision: Omit. The revision field is deprecated
The build-log.txt file is the stdout of the optional-operators-subscribe shell script. We should collect the stdout and save it to the file.
Both files should be saved inside the ARTIFACT_DIR for the role.